+++ docs/source/conf.py
@@ -35,8 +35,6 @@ autosummary_generate = True
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
 html_theme = "sphinx_rtd_theme"
-html_static_path = ["_static"]
-
 highlight_language = "python3"
 
 github_username = "adrianmrit"